url rewriting

+++ assets/php/functions.php
@@ -86,11 +86,15 @@ function convert_array_files($file) {
     return $result;
 }
 
-function valid_input_data($text_from,$email_to,$subject,$description) {
-    $result = array("val"=>true,"email_test"=>true,"from_test"=>true,"subject_test"=>true,"description_test"=>true);
-    if(preg_match('#^[^\s]+\@[^\s]+$#',trim($email_to))!==1) {
+function valid_input_data($text_from,$email_to,$email_from,$subject,$description) {
+    $result = array("val"=>true,"email_to_test"=>true,"email_from_test"=>true,"name_from_test"=>true,"subject_test"=>true,"description_test"=>true);
+    if(preg_match('#^([^\s]+\@[^\s]+)(\s*[\,\;]\s*([^\s]+\@[^\s]+))*$#',trim($email_to))!==1) {
         $result["val"] = false;
-        $result["email_test"] = false;
+        $result["email_to_test"] = false;
+    }
+    if(preg_match('#^([^\s]+\@[^\s]+)$#',trim($email_from))!==1) { //'#^([^\s]+\@[^\s]+)(\s*[\,\;]\s*([^\s]+\@[^\s]+))*$#'
+        $result["val"] = false;
+        $result["email_from_test"] = false;
     }
     /*if(preg_match('//',$text_from)!==1) {
         $result["val"] = false;
@@ -178,7 +182,7 @@ function clean_bdd($token="",$archive="") {
     }
     else {
         $date = new DateTime("now",new DateTimeZone("Europe/Paris")); // H:i:s
-        $result=mysql_get('SELECT a.* FROM archive WHERE "'.$date->format("Y-m-d").'" NOT BETWEEN DATE_FORMAT(send_date,"%Y-%m-%d") AND DATE_ADD(DATE_FORMAT(send_date,"%Y-%m-%d"), INTERVAL 10 DAY);');
+        $result=mysql_get('SELECT a.* FROM archive WHERE "'.$date->format("Y-m-d").'" NOT BETWEEN DATE_FORMAT(send_to_date,"%Y-%m-%d") AND DATE_ADD(DATE_FORMAT(send_to_date,"%Y-%m-%d"), INTERVAL 10 DAY);');
         if(!empty($result)) {
             $user_id_inline = "";
             foreach($result as $line) {
@@ -201,22 +205,44 @@ function clean_bdd($token="",$archive="") {
     }
 }
 
-function get_email_body($zipname,$token,$nb_files,$text_from,$subject,$description) {
+function get_email_to_body($zipname,$token,$nb_files,$text_from,$subject,$description) {
+    $zipname = preg_replace('#\.[zZ][iI][pP]$#','',$zipname);
     if(empty($text_from)) { $text_from='Quelqu\'un'; }
     if(empty($subject)) { $text_subject='n\'ayant pas d\'objet'; } else $text_subject='ayant pour objet <b>'.$subject.'</b>';
     $contenu_html='<p style="text-align:center;cursor:default;margin-top:20px;">Bonjour,</p>';
-    $contenu_html.='<p style="text-align:center;cursor:default;"><b>'.$text_from.'</b> vous a préparer une archive ZIP à téléchargé contenant '.$nb_files.' fichier(s) sur <a href="http://localhost/WeTransferCustom/">WeTransferCustom</a>';
+    $contenu_html.='<p style="text-align:center;cursor:default;"><b>'.$text_from.'</b> vous a préparer une archive ZIP à téléchargé contenant '.$nb_files.' fichier(s) sur <a target="_blank" href="http://localhost/WeTransferCustom/">WeTransferCustom</a>';
     $contenu_html.=' '.$text_subject.'.</p><p style="text-align:center;cursor:default;color:red;">Votre archive sera accesible pendant 10 jours à partir de cette date.</p>';
     if(!empty($description)) { $contenu_html.='<p style="text-align:center;cursor:default;">Message de <b>'.$text_from.'</b> :</p><p style="text-align:center;cursor:default;"><b>'.$description.'</b></p>'; }
-    $contenu_html.='<p style="text-align:center;cursor:default;;margin-top:50px;">Pour télécharger l\'archive, <a href="http://localhost/WeTransferCustom/download.php?archive='.$zipname.'&access='.$token.'">cliquez ici</a>.</p>';
-    $contenu_html.='<p style="text-align:center;cursor:default;">Pour supprimer l\'archive, <a href="http://localhost/WeTransferCustom/download.php?archive='.$zipname.'&access='.$token.'&delete=1">cliquez ici</a>.</p>';
+    //$contenu_html.='<p style="text-align:center;cursor:default;margin-top:50px;">Pour télécharger l\'archive, <a target="_blank" href="http://localhost/WeTransferCustom/download.php?archive='.$zipname.'&access='.$token.'">cliquez ici</a>.</p>';
+    $contenu_html.='<p style="text-align:center;cursor:default;margin-top:50px;">Pour télécharger l\'archive, <a target="_blank" href="http://localhost/WeTransferCustom/download/'.$token.'/'.$zipname.'">cliquez ici</a>.</p>';
+    //$contenu_html.='<p style="text-align:center;cursor:default;">Pour supprimer l\'archive, <a target="_blank" href="http://localhost/WeTransferCustom/download.php?archive='.$zipname.'&access='.$token.'&delete=1">cliquez ici</a>.</p>';
+    $contenu_html.='<p style="text-align:center;cursor:default;margin-top:40px;">Copyright WeTransferCustom 2020</p>';
+    return $contenu_html;
+}
+
+function get_email_from_body($zipname,$token,$nb_files,$text_from,$email_to_text,$subject,$description) {
+    $zipname = preg_replace('#\.[zZ][iI][pP]$#','',$zipname);
+    $contenu_html='<p style="text-align:center;cursor:default;margin-top:20px;">Bonjour,</p>';
+    $contenu_html.='<p style="text-align:center;cursor:default;">Nous vous remercions d\'avoir chosi <a target="_blank" href="http://localhost/WeTransferCustom/">WeTransferCustom.com</a> pour transférer vos fichiers en toute sécuriter.</p>';
+    $contenu_html.='<p style="text-align:center;cursor:default;margin-top:40px;">Récapitulatif de l\'envoi :</p>';
+    $contenu_html.='<div style="display:flex;cursor:default;margin-top:5px;flex-flow:row wrap;justify-content:center;align-items:center;align-content:stretch;">'; //<table style="flex: 0 1 auto;align-self: auto;order: 0;">    <tbody><tr style="padding: 4px;"><td style="padding: 4px;">Nombre de fichier dans l'archive</td><td style="padding: 4px;">1</td></tr>    <tr style="padding: 4px;"><td style="padding: 4px;">Destinataire(s)</td><td style="padding: 4px;">machin@chose.fr</td></tr></tbody></table></div>
+    $contenu_html.='    <table style="background-color:#EFEFEF;flex:0 1 auto;align-self:auto;order:0;">';
+    $contenu_html.='        <tr style="padding:4px;"><td style="padding:4px;">Nombre de fichier dans l\'archive</td><td style="padding:4px;color:blue;">'.$nb_files.'</td></tr>';
+    $contenu_html.='        <tr style="padding:4px;"><td style="padding:4px;">Destinataire(s)</td><td style="padding:4px;color:blue;">'.$email_to_text.'</td></tr>';
+    if(!empty(trim($text_from))) { $contenu_html.='     <tr style="padding:4px;"><td style="padding:4px;">Expéditeur</td><td style="padding:4px;color:blue;">'.$text_from.'</td></tr>'; }
+    if(!empty(trim($subject))) { $contenu_html.='       <tr style="padding:4px;"><td style="padding:4px;">Objet</td><td style="padding:4px;color:blue;">'.$subject.'</td></tr>'; }
+    if(!empty(trim($description))) { $contenu_html.='       <tr style="padding:4px;"><td style="padding:4px;">Description</td><td style="padding:4px;color:blue;">'.$description.'</td></tr>'; }
+    $contenu_html.='    </table>';
+    $contenu_html.='</div>';
+    //$contenu_html.='<p style="text-align:center;cursor:default;margin-top:50px;">Pour supprimer l\'archive, <a target="_blank" href="http://localhost/WeTransferCustom/download.php?archive='.$zipname.'&access='.$token.'&delete=1">cliquez ici</a>.</p>';
+    $contenu_html.='<p style="text-align:center;cursor:default;margin-top:50px;">Pour supprimer l\'archive, <a target="_blank" href="http://localhost/WeTransferCustom/download/'.$token.'/'.$zipname.'/delete">cliquez ici</a>.</p>';
     $contenu_html.='<p style="text-align:center;cursor:default;margin-top:40px;">Copyright WeTransferCustom 2020</p>';
     return $contenu_html;
 }
 
 function check_if_send() {
     $date = new DateTime("now",new DateTimeZone("Europe/Paris")); // H:i:s
-    $result=mysql_get('SELECT d.*, u.token FROM archive a INNER JOIN user u WHERE send=0 OR send_date IS NULL;');
+    $result=mysql_get('SELECT d.*, u.token FROM archive a INNER JOIN user u WHERE send_to=0 OR send_to_date IS NULL OR send_from=0 OR send_from_date IS NULL;');
     if(!empty($result)) {
         $test = false;
         foreach($result as $archive) {
@@ -228,23 +254,37 @@ function check_if_send() {
             }
             if($test) {
                 $token = $archive["token"];
-                $email_to_array = preg_split('#\,#',$archive["email_to"],0,PREG_SPLIT_NO_EMPTY);
-                $text_from = $archive["email_from"];
+                $text_from = $archive["name_from"];
                 $subject = $archive["subject"];
                 $description = $archive["description"];
                 $nb_files = $archive["nb_files"];
-                //if(empty($text_from)) { $text_from='Quelqu\'un'; }
-                /*$contenu_html='<p>Bonjour,<br><br><b>'.$text_from.'</b> vous a préparer une archive ZIP contenant '.$nb_files.' fichier(s) à téléchargé sur <a href="http://127.0.0.1/wetransfer_custom/">WeTransferCustom</a>';
-                $contenu_html.=' ayant pour objet <b>'.$subject.'</b>.</p><br><p><span style="color:red;">Votre archive sera accesible pendant 10 jours à partir de la date de cet email.</span></p><br><br>';
-                if(!empty($description)) { $contenu_html.='<p>Message de <b>'.$text_from.'</b> :</p><br><p><b>'.$description.'</b></p><br><br>'; }
-                $contenu_html.='<p style="text-align:center;">Pour télécharger l\'archive, <a href="http://127.0.0.1/wetransfer_custom/download.php?archive='.$zipname.'&access='.$token.'">cliquez ici</a>.</p><br><br>';
-                $contenu_html.='<p style="text-align:center;">Pour supprimer l\'archive, <a href="http://127.0.0.1/wetransfer_custom/download.php?archive='.$zipname.'&access='.$token.'&delete=1">cliquez ici</a>.</p>';*/
-                $contenu_html=get_email_body($zipname,$token,$nb_files,$text_from,$subject,$description);
-                //$contenu_text=''; //'Bonjour,\n\n'.$text_from.' vous a préparer une archive ZIP à téléchargé sur WeTransferCustom.';
-                $result_send_mail = send_mail("Transfert de fichier(s) - WeTransferCustom",$contenu_html,$email_to_array); //,$contenu_text
-                if(!empty($result_send_mail["val"])) {
-                    $date = new DateTime("now",new DateTimeZone("Europe/Paris"));
-                    $update = mysql_set('UPDATE archive SET send=1, send_date = "'.$date->format("Y-m-d H:i:s").'"  WHERE id = '.$archive["id"].' ;');
+                $email_from_text = $archive["email_from"];
+                //
+                if(empty($archive["send_to_date"]) || empty($archive["send_to"])) {
+                    $email_to_array = preg_split('#\,#',$archive["email_to"],0,PREG_SPLIT_NO_EMPTY);
+                    //if(empty($text_from)) { $text_from='Quelqu\'un'; }
+                    /*$contenu_html='<p>Bonjour,<br><br><b>'.$text_from.'</b> vous a préparer une archive ZIP contenant '.$nb_files.' fichier(s) à téléchargé sur <a href="http://127.0.0.1/wetransfer_custom/">WeTransferCustom</a>';
+                    $contenu_html.=' ayant pour objet <b>'.$subject.'</b>.</p><br><p><span style="color:red;">Votre archive sera accesible pendant 10 jours à partir de la date de cet email.</span></p><br><br>';
+                    if(!empty($description)) { $contenu_html.='<p>Message de <b>'.$text_from.'</b> :</p><br><p><b>'.$description.'</b></p><br><br>'; }
+                    $contenu_html.='<p style="text-align:center;">Pour télécharger l\'archive, <a href="http://127.0.0.1/wetransfer_custom/download.php?archive='.$zipname.'&access='.$token.'">cliquez ici</a>.</p><br><br>';
+                    $contenu_html.='<p style="text-align:center;">Pour supprimer l\'archive, <a href="http://127.0.0.1/wetransfer_custom/download.php?archive='.$zipname.'&access='.$token.'&delete=1">cliquez ici</a>.</p>';*/
+                    $contenu_html=get_email_to_body($zipname,$token,$nb_files,$text_from,$subject,$description);
+                    //$contenu_text=''; //'Bonjour,\n\n'.$text_from.' vous a préparer une archive ZIP à téléchargé sur WeTransferCustom.';
+                    $result_send_to_mail = send_mail("Transfert de fichier(s) - WeTransferCustom",$contenu_html,$email_to_array); //,$contenu_text
+                    if(!empty($result_send_to_mail["val"])) {
+                        $dateA = new DateTime("now",new DateTimeZone("Europe/Paris"));
+                        $update = mysql_set('UPDATE archive SET send_to=1, send_to_date = "'.$dateA->format("Y-m-d H:i:s").'"  WHERE id = '.$archive["id"].' ;');
+                    }
+                }
+                //
+                if(empty($archive["send_from_date"]) || empty($archive["send_from"])) {
+                    $email_from_array = preg_split('#\s*[\,\;]\s*#',$email_from,0,PREG_SPLIT_NO_EMPTY);
+                    $contenu_html=get_email_from_body($zipname,$token,$nb_files,$text_from,$email_from_text,$subject,$description);
+                    $result_send_from_mail = send_mail("WeTransferCustom - Information",$contenu_html,$email_from_array); //,$contenu_text
+                    if(!empty($result_send_from_mail["val"])) {
+                        $dateB = new DateTime("now",new DateTimeZone("Europe/Paris"));
+                        $update = mysql_set('UPDATE archive SET send_from=1, send_from_date = "'.$dateB->format("Y-m-d H:i:s").'"  WHERE id = '.$archive["id"].' ;');
+                    }
                 }
             }
         }
